Move LWRPs to custom_resources - Sketch out two new resources, `opsview_client` & `opsview_client_rest` one to manage client install and one to manage it' registration in the console.
