+++ Guardfile
@@ -27,13 +27,3 @@ guard :rspec, cmd: 'chef exec rspec --color' do
   watch(%r{^files/(.+)})
   watch(%r{^templates/(.+)})
 end
-
-# guard 'kitchen' do
-#  watch(%r{test/.+})
-#  watch(%r{^recipes/(.+)\.rb$})
-#  watch(%r{^attributes/(.+)\.rb$})
-#  watch(%r{^files/(.+)})
-#  watch(%r{^templates/(.+)})
-#  watch(%r{^providers/(.+)\.rb})
-#  watch(%r{^resources/(.+)\.rb})
-# end
