+++ templates/default/NSC.ini.erb
@@ -29,7 +29,7 @@ CheckExternalScripts.dll
 [Settings]
 ;# OBFUSCATED PASSWORD
 ;  This is the same as the password option but here you can store the password in an obfuscated manner.
-;  *NOTICE* obfuscation is *NOT* the same as encryption, someone with access to this file can still figure out the 
+;  *NOTICE* obfuscation is *NOT* the same as encryption, someone with access to this file can still figure out the
 ;  password. Its just a bit harder to do it at first glance.
 ;obfuscated_password=Jw0KAUUdXlAAUwASDAAB
 ;
@@ -44,7 +44,7 @@ CheckExternalScripts.dll
 ;allowed_hosts=127.0.0.1/32
 ;
 ;# USE THIS FILE
-;  Use the INI file as opposed to the registry if this is 0 and the use_reg in the registry is set to 1 
+;  Use the INI file as opposed to the registry if this is 0 and the use_reg in the registry is set to 1
 ;  the registry will be used instead.
 use_file=1
 ;
@@ -144,7 +144,7 @@ ps1=cmd /c echo %SCRIPT% %ARGS%; exit($lastexitcode) | powershell.exe -command -
 
 [Check System]
 ;# CPU BUFFER SIZE
-;  Can be anything ranging from 1s (for 1 second) to 10w for 10 weeks. Notice that a larger buffer will waste memory 
+;  Can be anything ranging from 1s (for 1 second) to 10w for 10 weeks. Notice that a larger buffer will waste memory
 ;  so don't use a larger buffer then you need (ie. the longest check you do +1).
 ;CPUBufferSize=1h
 ;
@@ -209,12 +209,12 @@ custom.ini
 ;interval=5
 ;
 ;# ENCRYPTION METHOD
-;   This option determines the method by which the send_nsca client will encrypt the packets it sends 
-;   to the nsca daemon. The encryption method you choose will be a balance between security and 
+;   This option determines the method by which the send_nsca client will encrypt the packets it sends
+;   to the nsca daemon. The encryption method you choose will be a balance between security and
 ;   performance, as strong encryption methods consume more processor resources.
 ;   You should evaluate your security needs when choosing an encryption method.
 ;
-; Note: The encryption method you specify here must match the decryption method the nsca daemon uses 
+; Note: The encryption method you specify here must match the decryption method the nsca daemon uses
 ;       (as specified in the nsca.cfg file)!!
 ; Values:
 ;	0 = None	(Do NOT use this option)
@@ -231,13 +231,13 @@ custom.ini
 ;encryption_method=14
 ;
 ;# ENCRYPTION PASSWORD
-;  This is the password/passphrase that should be used to encrypt the sent packets. 
+;  This is the password/passphrase that should be used to encrypt the sent packets.
 ;password=
 ;
 ;# BIND TO ADDRESS
 ;  Allows you to bind server to a specific local address. This has to be a dotted ip adress not a hostname.
 ;  Leaving this blank will bind to "one" local interface.
-; -- not supported as of now -- 
+; -- not supported as of now --
 ;bind_to_address=
 ;
 ;# LOCAL HOST NAME
